+++ distance_sensor.h
@@ -41,7 +41,7 @@ public:
         _current_value = 0;
         _previous_value = 0;
         _offset_value = _min_value;  //< オフセット値は前回の周期の物を使う
-        _min_value = UINT32_MAX;
+        _min_value = UINT16_MAX;
         while (_previous_value < _current_value) {
             _previous_value = _current_value;
             _current_value = _photo_transistor.read();
@@ -53,7 +53,7 @@ public:
      * @fn _offset の値のために、任意のタイマ割り込みで呼ぶ
      */
     inline void interruptSamplingMinValue() {
-        const uint32_t temp_value = _photo_transistor.read();
+        const uint16_t temp_value = _photo_transistor.read();
         if (temp_value < _min_value) _min_value = temp_value;
     }
 
@@ -62,7 +62,7 @@ public:
     /**
      * @param Charge capacitor (can't set us unit).
      */
-    uint32_t read() const {
+    uint16_t read() const {
         while(!_get_flag){} //< 最新の値を持ってくるまで待つ
         return _previous_value - _offset_value; //< ここで一旦値を保存して getDistance_mm を呼ぶのがいいかも。
 //        return _photo_transistor.read();
@@ -70,19 +70,19 @@ public:
 
 private:
 
-    uint32_t getDistance_mm();
+    uint16_t getDistance_mm();
 
-    uint32_t convert_12bit_to_mm(uint16_t value);
+    uint16_t convert_12bit_to_mm(uint16_t value);
 
 
     PWMOut _led;
     AnalogInDMAStream _photo_transistor;
     TIM_HandleTypeDef &_sampling_htim_x;
-    uint32_t _value;          //< 最終的な距離に直す値
-    uint32_t _current_value;  //< _get_flag のための微分する最新の (値現在周期の値)
-    uint32_t _previous_value; //< _get_flag のための微分する１つ前の値 (値現在周期の値)
-    uint32_t _min_value;      //< 最低値 (値現在周期の値)
-    uint32_t _offset_value;   //< 1つ前の周期の最低値を記録したもの(オフセットとして使う) (1つ前の周期の値)
+    uint16_t _value;          //< 最終的な距離に直す値
+    uint16_t _current_value;  //< _get_flag のための微分する最新の (値現在周期の値)
+    uint16_t _previous_value; //< _get_flag のための微分する１つ前の値 (値現在周期の値)
+    uint16_t _min_value;      //< 最低値 (値現在周期の値)
+    uint16_t _offset_value;   //< 1つ前の周期の最低値を記録したもの(オフセットとして使う) (1つ前の周期の値)
 
     bool _get_flag;
 };
